blog_agent/blog_agent/requirements.txt
langchain==0.3.25
langchain-core==0.3.60
langchain-openai==0.3.16
langgraph==0.4.1
streamlit==1.45.0
httpx==0.28.1
pydantic==2.11.4
python-dotenv==1.1.0
